refactor(core): update layout
