+++ cmd/cdnApp.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/ebrahimahmadi/ar-cli/internals/pkg/utils"
+	"github.com/ebrahimahmadi/ar-cli/internals/utils"
 	"github.com/ebrahimahmadi/ar-cli/pkg/api"
 	"github.com/ebrahimahmadi/ar-cli/pkg/helpers"
 	"github.com/ebrahimahmadi/ar-cli/pkg/validator"
